ViewとControllerのコードを整理
