+++ app/assets/stylesheets/header.scss
@@ -7,7 +7,7 @@
   @include container;
   display: flex;
   align-items: center;
-  justify-content: flex-end;
+  justify-content: space-between;
   height: 60px;
   width: 100%;
   position: fixed;
