+++ app/controllers/articles_controller.rb
@@ -1,6 +1,9 @@
 class ArticlesController < ApplicationController
   before_action :authenticate_user!
   before_action :set_categories
+  before_action :set_article, only: [:update, :destroy, :article_item_destroy]
+  before_action :set_article_questions, only: [:show, :edit]
+  before_action :browser_rimit, only: [:show, :edit]
 
   def index
     @articles = current_user.articles.all.page(params[:page]).per(10)
@@ -8,18 +11,14 @@ class ArticlesController < ApplicationController
 
   def new
     @article = Article.new
-    @article_item = ArticleItem.new
-    @article_questions = current_user.article_questions
   end
 
   def create
     @article = current_user.articles.build(article_params)
 
     if @article.save
-      params[:items].each do |item|
-        @article_item = @article.article_items.build(article_question_id: item[1][0], body: item[1][1])
-        @article_item.save!
-      end
+      article_item_create
+
       flash[:notice] = "新規投稿しました！"
       redirect_to articles_path
     else
@@ -30,43 +29,18 @@ class ArticlesController < ApplicationController
   end
 
   def show
-    @article = Article.find(params[:id])
-    @article_questions = current_user.article_questions
-
-    if @article.user != current_user
-      flash[:alert] = "他のユーザーの投稿を見ることはできません！"
-      redirect_to articles_path
-    end
   end
 
   def edit
-    @article = Article.find(params[:id])
     @article_items = @article.article_items
-    @article_questions = current_user.article_questions
-
-    if @article.user != current_user
-      flash[:alert] = "他のユーザーの投稿は編集できません！"
-      redirect_to articles_path
-    end
   end
 
   def update
-    @article = Article.find(params[:id])
     @article_items = @article.article_items
 
     if @article.update(update_article_params)
-      # if params.include?("items") タイトルだけのメモの編集でタイトルだけを編集した場合に出るエラー対策
-      if params.include?("items")
-        i = 0
-        params[:items].each do |item|
-          if @article_items.length < i + 1
-            @article_items.create!(article_question_id: item[1][0], body: item[1][1])
-          else
-            @article_items[i].update!(article_question_id: item[1][0], body: item[1][1])
-          end
-          i += 1
-        end
-      end
+      article_item_edit
+
       flash[:notice] = "投稿を編集しました！"
       redirect_to articles_path
     else
@@ -75,15 +49,12 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    @article = Article.find(params[:id])
-
     @article.destroy!
     flash[:alert] = "投稿を削除しました！"
     redirect_to articles_path
   end
 
   def article_item_destroy
-    @article = Article.find(params[:id])
     @article_items = @article.article_items
 
     @article_items.last.destroy!
@@ -91,10 +62,46 @@ class ArticlesController < ApplicationController
 
   def search
     @search_word = params[:search_word]
-    @articles = current_user.articles.search(@search_word)
+    @articles = current_user.articles.search(@search_word).all.page(params[:page]).per(10)
   end
 
   private
+
+    def article_item_create
+      params[:items].each do |item|
+        @article_item = @article.article_items.build(article_question_id: item[1][0], body: item[1][1])
+        @article_item.save!
+      end
+    end
+
+    def article_item_edit
+      i = 0
+      params[:items].each do |item|
+        if @article_items.length < i + 1
+          @article_items.create!(article_question_id: item[1][0], body: item[1][1])
+        else
+          @article_items[i].update!(article_question_id: item[1][0], body: item[1][1])
+        end
+        i += 1
+      end
+    end
+
+    def browser_rimit
+      @article = Article.find(params[:id])
+
+      if @article.user != current_user
+        flash[:alert] = "他のユーザーの投稿の閲覧・編集はできません！"
+        redirect_to articles_path
+      end
+    end
+
+    def set_article_questions
+      @article_questions = current_user.article_questions
+    end
+
+    def set_article
+      @article = current_user.articles.find(params[:id])
+    end
 
     def article_params
       params.require(:article).permit(:title, :category_id, :user_id)
