--- app/helpers/article_lists_helper.rb
@@ -1,2 +1,0 @@
-module ArticleListsHelper
-end
