--- app/helpers/article_questions_helper.rb
@@ -1,2 +1,0 @@
-module ArticleQuestionsHelper
-end
