--- app/helpers/favorites_helper.rb
@@ -1,2 +1,0 @@
-module FavoritesHelper
-end
