--- app/helpers/home_helper.rb
@@ -1,2 +1,0 @@
-module HomeHelper
-end
