+++ app/views/favorites/index.html.erb
@@ -1,36 +1,4 @@
 <main class="main">
-  <h2 class="mb-4">お気に入り一覧</h2>
-  <% @articles.each do |article| %>
-    <ul class="list-group">
-      <li class="list-group-item">
-        <ul>
-          <li><%= link_to(article.title, article_path(article))%></li>
-        </ul>
-        <ul>
-          <li>
-            <% if article.favorite == false %>
-              <%= link_to(article_favorites_path(article), remote: true, method: :patch) do %>
-                <i class="far fa-star fa-2x"></i>
-              <% end %>
-            <% else %>
-              <%= link_to(article_favorites_path(article), remote: true, method: :patch) do %>
-                <i class="fas fa-star fa-2x"></i>
-              <% end %>
-            <% end %>
-          </li>
-          <li>
-            <%= link_to(edit_article_path(article)) do %>
-              <i class="fas fa-edit fa-2x"></i>
-            <% end %>
-          </li>
-          <li>
-            <%= link_to(article_path(article),method: :delete, class: "destroy", data: {confirm: "投稿を削除しますか？"}, class: "destroy") do %>
-              <i class="fas fa-trash fa-2x"></i>
-            <% end %>
-          </li>
-        </ul>
-      </li>
-    </ul>
-  <% end %>
+  <%= render partial: "shared/article_index", locals: { articles: @articles, subject: "お気に入り一覧"}%>
   <%= paginate @articles %>
 </main>
