+++ app/views/home/index.html.erb
@@ -44,4 +44,6 @@
   </div>
 </section>
 
-<%= render partial: "shared/footer" %>
+<footer class="footer">
+  <p>Dig Memo 2020</p>
+</footer>
