--- app/views/shared/_footer.html.erb
@@ -1,3 +1,0 @@
-<footer class="footer">
-  <p>Dig Memo 2020</p>
-</footer>
