+++ app/views/shared/_header.html.erb
@@ -1,9 +1,10 @@
 <header class="header">
-  <nav>
-    <ul class="header__list">
-      <li class="header__list__item"><%= link_to("新規登録",new_user_registration_path ) %></li>
-      <li class="header__list__item"><%= link_to("ログイン",new_user_session_path ) %></li>
-      <li class="header__list__item"><%= link_to("ゲストログイン（閲覧用）",users_guest_sign_in_path, method: :post ) %></li>
-    </ul>
-  </nav>
+  <%= link_to(root_path) do %>
+    <img src="/assets/logo.png" alt=""/>
+  <% end %>
+  <ul class="header__list">
+    <li class="header__list__item"><%= link_to("新規登録",new_user_registration_path ) %></li>
+    <li class="header__list__item"><%= link_to("ログイン",new_user_session_path ) %></li>
+    <li class="header__list__item"><%= link_to("ゲストログイン（閲覧用）",users_guest_sign_in_path, method: :post ) %></li>
+  </ul>
 </header>
