--- config/locales/enum.yml
@@ -1,12 +1,0 @@
-ja:
-  enums:
-    article_item:
-      question:
-        why: "なぜ？"
-        how: "どのように？"
-        where: "どこで？"
-        what: "何？"
-        summary: "一言で言うと？"
-        application: "どうやって応用する？"
-        other: "その他"
-        who: "誰？"
