+++ config/routes.rb
@@ -1,8 +1,4 @@
 Rails.application.routes.draw do
-  # ゲストログイン
-  devise_scope :user do
-    post "/users/guest_sign_in", to: "users/sessions#new_guest"
-  end
   root "home#index"
   get "favorites/index"
   get "search", to: "articles#search", as: :search
@@ -14,14 +10,18 @@ Rails.application.routes.draw do
   resource :category, only: [:destroy]
   resources :categories, only: [:index, :new, :create, :edit]
   resources :article_questions, only: [:create]
+  resources :articles do
+    resource :favorites, only: [:index, :update]
+  end
   devise_for :users, controllers: {
     registrations: "users/registrations",
     sessions: "users/sessions",
     passwords: "users/passwords",
     confirmations: "users/confirmations",
   }
-  resources :articles do
-    resource :favorites, only: [:index, :update]
+  # ゲストログイン
+  devise_scope :user do
+    post "/users/guest_sign_in", to: "users/sessions#new_guest"
   end
 
   if Rails.env.development?
